backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "faster-whisper>=1.2.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langgraph>=0.6.6",
    "msal>=1.33.0",
    "pillow>=11.3.0",
    "pydantic>=2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "torch>=2.0.0",
    "uvicorn>=0.35.0",
]
